feat: 유저 검증 로직 추가

+++ src/main/java/com/swyp/playground/domain/comment/service/CommentService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.swyp.playground.domain.comment.domain.Comment;
 import com.swyp.playground.domain.comment.dto.WriteCommentDto;
 import com.swyp.playground.domain.comment.repository.CommentRepository;
+import com.swyp.playground.domain.parent.repository.ParentRepository;
 
 
 @Service
@@ -17,12 +18,13 @@ public class CommentService {
     @Autowired
     private CommentRepository commentRepository;
 
-    /* -- POST -- */
+    @Autowired
+    private ParentRepository parentRepository;
+
     public Comment writeComment(Comment comment) {
         return commentRepository.save(comment);
     }
 
-    /* -- GET -- */
     public List<Comment> getAllComments() {
         return commentRepository.findAll();
     }
@@ -35,14 +37,17 @@ public class CommentService {
         return commentRepository.findById(id);
     }
 
-    /* -- UPDATE -- */
-    public void patchCommentById(Long id, WriteCommentDto newComment) {
+    public void patchCommentById(Long id, WriteCommentDto newComment, String email) {
         try {
             Comment prevComment = commentRepository.findById(id)
                             .orElseThrow(() -> new NullPointerException("해당 댓글이 존재하지 않습니다."));
             
-            if (prevComment.getContent() != newComment.getContent()) {
-                prevComment.setContent(newComment.getContent());
+            if (prevComment.getWrittenBy() == parentRepository.findByEmail(email).get().getNickname()) {
+                if (prevComment.getContent() != newComment.getContent()) {
+                    prevComment.setContent(newComment.getContent());
+                }
+            } else {
+                throw new IllegalAccessError("작성한 유저만 수정할 수 있습니다.");
             }
             
             commentRepository.save(prevComment);
@@ -51,8 +56,16 @@ public class CommentService {
         }
     }
 
-    /* -- DELETE -- */
-    public void deleteComment(Long id) {
-        commentRepository.deleteById(id);
+    public void deleteComment(Long id, String user) {
+        try {
+            String prevUser = commentRepository.findById(id).get().getWrittenBy();
+            if (prevUser.equals(user)) {
+                commentRepository.deleteById(id);
+            } else {
+                throw new IllegalAccessError("작성한 유저만 삭제할 수 있습니다.");
+            }
+        } catch (Exception e) {
+            System.err.println(e);
+        }
     }
 }
